+++ samplesheets/src/components/PageHeader.vue
@@ -1,10 +1,13 @@
 <template>
-  <div class="row sodar-subtitle-container bg-white sticky-top">
+  <div class="row sodar-subtitle-container bg-white sticky-top"
+       id="sodar-ss-vue-subtitle">
     <h3><i class="fa fa-flask"></i> Sample Sheets</h3>
     <b-nav v-if="app.sheetsAvailable"
+           id="sodar-ss-nav-tabs"
            pills
            class="sodar-ss-nav ml-4 mr-auto">
       <b-nav-item v-for="(studyInfo, studyUuid, index) in app.sodarContext['studies']"
+                  :id="'sodar-ss-tab-study-' + studyUuid"
                   :key="index"
                   @mousedown="app.handleStudyNavigation(studyUuid)"
                   v-b-tooltip.hover
@@ -13,7 +16,8 @@
                   :disabled="!app.sheetsAvailable || app.gridsBusy">
         <i class="fa fa-list-alt"></i> {{ studyInfo['display_name'] | truncate(20) }}
       </b-nav-item>
-      <b-nav-item @mousedown="app.showOverview()"
+      <b-nav-item id="sodar-ss-tab-overview"
+                  @mousedown="app.showOverview()"
                   :active="app.overviewActive"
                   :disabled="!app.sheetsAvailable || app.gridsBusy">
         <i class="fa fa-sitemap"></i> Overview
@@ -32,6 +36,7 @@
               :key="index">
           <b-dropdown-item
               href="#"
+              :id="'sodar-ss-nav-study-' + studyUuid"
               class="sodar-ss-nav-item"
               @click="app.handleStudyNavigation(studyUuid)">
             <i class="fa fa-fw fa-list-alt text-info"></i> {{ studyInfo['display_name'] }}
@@ -40,6 +45,7 @@
               v-for="(assayInfo, assayUuid, assayIndex) in studyInfo['assays']"
               :key="assayIndex"
               href="#"
+              :id="'sodar-ss-nav-assay-' + assayUuid"
               class="sodar-ss-nav-item"
               @click="app.handleStudyNavigation(studyUuid, assayUuid)">
             <i class="fa fa-fw fa-table text-danger ml-4"></i> {{ assayInfo['display_name'] }}
@@ -47,13 +53,14 @@
         </span>
         <b-dropdown-item
           href="#"
+          id="sodar-ss-nav-overview"
           class="sodar-ss-nav-item"
           @click="app.showOverview()">
           <i class="fa fa-fw fa-sitemap"></i> Overview
         </b-dropdown-item>
       </b-dropdown>
       <b-dropdown
-          id="sodar-ss-buttons-op"
+          id="sodar-ss-op-dropdown"
           :disabled="app.gridsBusy"
           right
           variant="primary"
@@ -61,12 +68,14 @@
         <b-dropdown-item
             v-if="!app.sheetsAvailable &&
                   app.sodarContext['perms']['edit_sheet']"
+            class="sodar-ss-op-item"
             :href="'import/' + app.projectUuid">
           <i class="fa fa-fw fa-upload"></i> Import ISAtab
         </b-dropdown-item>
         <b-dropdown-item
             v-if="app.sheetsAvailable &&
                   app.sodarContext['perms']['edit_sheet']"
+            class="sodar-ss-op-item"
             :href="'import/' + app.projectUuid">
           <i class="fa fa-fw fa-refresh"></i> Replace ISAtab
         </b-dropdown-item>
@@ -74,6 +83,7 @@
             v-if="app.sheetsAvailable &&
                   !app.renderError &&
                   app.sodarContext['perms']['create_dirs']"
+            class="sodar-ss-op-item"
             :href="'dirs/' + app.projectUuid">
           <i class="fa fa-fw fa-database"></i>
           <span v-if="app.sodarContext['irods_status']">Update</span><span v-else>Create</span> iRODS Directories
@@ -81,6 +91,7 @@
         <b-dropdown-item
             v-if="app.sheetsAvailable &&
                   app.sodarContext['perms']['delete_sheet']"
+            class="sodar-ss-op-item"
             :href="'delete/' + app.projectUuid">
           <i class="fa fa-fw fa-close"></i> Delete Sheets and Data
         </b-dropdown-item>