+++ CHANGELOG.rst
@@ -44,6 +44,7 @@ Changed
     - Redesign study apps to work with Vue.js viewer (#436)
     - Display study shortcuts as link column instead of separate table (#464)
     - Do not display shortcuts in cancer study app for mass spectrometry assays (workaround for #482)
+    - Move ``get_material_count()`` from views into Investigation model
 
 Fixed
 -----